Fixed Dates
Run the FixTheDataBasedOnNearestKnown after downloading the demand data, this will output a folder containing the new fixed data.
